+++ src/gui/npcdialog.cpp
@@ -84,7 +84,7 @@ NpcDialog::NpcDialog(int npcId) :
     // Setup output text box
     mTextBox = new BrowserBox(BrowserBox::AUTO_WRAP);
     mTextBox->setOpaque(false);
-    mTextBox->setMaxRow(static_cast<int>(config.getIntValue("ChatLogLength")));
+    mTextBox->setMaxRow(config.getIntValue("ChatLogLength"));
     mTextBox->setLinkHandler(mItemLinkHandler);
 
     mScrollArea = new ScrollArea(mTextBox);
