+++ src/gui/widgets/chattab.cpp
@@ -60,8 +60,7 @@ ChatTab::ChatTab(const std::string &name) :
 
     mTextOutput = new BrowserBox(BrowserBox::AUTO_WRAP);
     mTextOutput->setOpaque(false);
-    mTextOutput->setMaxRow(static_cast<int>(
-        config.getIntValue("ChatLogLength")));
+    mTextOutput->setMaxRow(config.getIntValue("ChatLogLength"));
     if (chatWindow)
         mTextOutput->setLinkHandler(chatWindow->mItemLinkHandler);
     mTextOutput->setAlwaysUpdate(false);
